Fixed some warnings

+++ src/interval.cpp
@@ -4,14 +4,14 @@ Interval::Interval(double milliseconds, py::function function) : milliseconds(mi
 	nextTime = alt::ICore::Instance().GetNetTime() + (long)milliseconds;
 }
 
-void Interval::TimeWarning(long time, const std::string& resourceName) {
+void Interval::TimeWarning(uint32_t time, const std::string& resourceName) {
 	auto inspect = py::module_::import("inspect");
 	auto path = inspect.attr("getfile")(function).cast<std::string>();
 	std::string str = std::string("Interval at ").append(resourceName).append(":").append(path.substr(path.find_last_of("/\\") + 1)).append(" at function ").append(function.attr("__name__").cast<std::string>()).append(" was too long ").append(std::to_string(alt::ICore::Instance().GetNetTime() - time)).append("ms");
 	alt::ICore::Instance().LogWarning(str);
 }
 
-bool Interval::Update(long time) {
+bool Interval::Update(uint32_t time) {
 	if (time >= nextTime && running) {
 		try {
 			function();
